refactor: extract _shared.py from phi_scan/report package
